+++ src/Engine/asset_loaders/audio_resources.rs
@@ -1,3 +1,5 @@
+// Gère les assets audios.
+
 use std::collections::HashMap;
 
 use bevy::prelude::*;
