+++ src/Engine/asset_loaders/graphic_resources.rs
@@ -7,8 +7,7 @@ use bevy::prelude::*;
 #[derive(Resource, Debug, Default)]
 pub struct GraphicsAssets {
     pub logo: Handle<Image>,
-    pub ascii_sheet_layout: Handle<TextureAtlasLayout>,
-    pub ascii_sheet_img: Handle<Image>,
+    pub logo_layout: Handle<TextureAtlasLayout>,
     pub font: Handle<Font>,
     pub textures: HashMap<&'static str, Handle<Image>>,
     pub map_textures: HashMap<&'static str, Handle<Image>>,
